requirements.txt
altgraph==0.17.2
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
future==0.18.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
macholib==1.15.2
mysql-connector-python==8.1.0
Pillow==9.5.0
protobuf==4.21.12
pydantic==2.3.0
pydantic_core==2.6.3
pyftpdlib==1.5.7
python-decouple==3.8
python-multipart==0.0.6
six==1.15.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.7.0
urllib3==2.0.5
uvicorn==0.22.0
